+++ SistemaDeGarage/src/main/java/org/example/Camion.java
@@ -3,7 +3,7 @@ package org.example;
 public class Camion extends Vehiculo{
 
 
-    public Camion(String modelo, String marca, String patente, double horasEstimadas) {
+    public Camion(String modelo, String marca, String patente, int horasEstimadas) {
         super(modelo, marca, patente, horasEstimadas);
     }
